refactor(fe): extract trigger and problem list from AssignmentAccordionItem
